i think i fixed the factory constructor
The factory constructor works. You don't need to put anything at the start in the constructor. we can push the stuff in the game model. you just need a method to put the stuff that is given as an array or vector into the existing 2dArray. In the constructor the only thing u need to do is initialise the 2d vector variable